requirements.txt
aiogram==3.1.1
aiosqlite
aiohttp
python-dotenv
loguru
pillow
